Update build process for proper compilation on Windows
 ! Windows compilation has only been tested on Windows 10 and no support will
   be offered
 * Updated Windows library linking in CMakeLists.txt
 * Updated README.md with new Windows compilation instructions
 * Renamed the following files and their associated headers for easier
   compilation with Visual Studio:
        bee/core/room.cpp to bee/core/rooms.cpp
        bee/util/network.cpp to bee/util/networking.cpp
        bee/util/sound.cpp to bee/util/soundeffects.cpp
        ! This was necessary because Visual Studio is not good at compiling
          files with the same name such as bee/resources/room.cpp
        ! I didn't initially want to make this change but it makes the files a
          little more descriptive hopefully
 * Added a bunch of static_casts and specific type conversions to stop Visual
   Studio from complaining about possible loss of data for deliberate truncated
   type conversions
 * Fixed incorrect class vs. struct forward declarations
 * Fixed a bug with map iteration in bee::messenger::clear_log()
 * Fixed scale calculations in PhysicsBody::set_shape()
 * Changed Particle coordinates from ints to doubles

+++ bee/core/console.cpp
@@ -30,7 +30,7 @@
 #include "input.hpp"
 #include "keybind.hpp"
 #include "resources.hpp"
-#include "room.hpp"
+#include "rooms.hpp"
 
 #include "../render/drawing.hpp"
 #include "../render/renderer.hpp"
@@ -430,8 +430,8 @@ namespace bee{ namespace console {
 	*/
 	int internal::draw() {
 		// Get the view offset
-		int cx = ui_handle->get_corner_x();
-		int cy = ui_handle->get_corner_y();
+		int cx = static_cast<int>(ui_handle->get_corner_x());
+		int cy = static_cast<int>(ui_handle->get_corner_y());
 
 		ui_text_entry->set_corner_x(cx);
 		ui_text_entry->set_corner_y(cy+rect.h);
@@ -455,7 +455,9 @@ namespace bee{ namespace console {
 		// Split lines if they are wider than the console window
 		for (auto it=lines.begin(); it!=lines.end(); ++it) {
 			if (engine->font_default->get_string_width(*it) > rect.w) {
-				int c = 1.5*rect.w/engine->font_default->get_font_size()-1;
+				int c = static_cast<int>(
+					1.5 * rect.w / engine->font_default->get_font_size()
+				) - 1;
 
 				lines.emplace(it+1, (*it).substr(c));
 				*it = (*it).substr(0, c);
