+++ bee/render/particle/changer.cpp
@@ -96,7 +96,7 @@ namespace bee {
 		double cx = get_following_x(system_x);
 		double cy = get_following_y(system_y);
 
-		return draw_rectangle(cx+x, cy+y, w, h, 1, get_enum_color(color));
+		return draw_rectangle(static_cast<int>(cx+x), static_cast<int>(cy+y), w, h, 1, get_enum_color(color));
 	}
 }
